requirements.txt
google-adk
requests
playwright
gspread
google-auth
ollama

